refactor: 코드리뷰 반영(필요없는 주석 삭제, import 경로 수정, empty view 컴포넌트화)
