+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 import useEasyNavigate from './useEasyNavigate';
+import useToast from './useToast';
 
-export { useEasyNavigate };
+export { useEasyNavigate, useToast };
