+++ src/pages/host/components/index.ts
@@ -1,4 +1,5 @@
 import HostMyClassCard from './HostMyClassCard/HostMyClassCard';
 import ApplicantListModal from './ApplicantListModal/ApplicantListModal';
+import ClassManageEmptyView from './ClassManageEmptyView/ClassManageEmptyView';
 
-export { HostMyClassCard, ApplicantListModal };
+export { HostMyClassCard, ApplicantListModal, ClassManageEmptyView };
